ft: Calculate height from bounding box
In alacritty/alacritty#2693, we see two problems: powerline separator
glyphs are offset both vertically and horizontally.  Having meticulously
redesigned the powerline separator glyphs for Hack in
source-foundry/Hack#234, I can say with some degree of confidence that
the horizontal misalignment is due to the separator glyphs not being
designed to be larger than the bounding box.  That is to say, it's a
font problem.

The vertical misaligment, however, comes down to how this library is
calculating the font height.  FreeType provides a height metric, but it
represents "the vertical distance between two consecutive baselines" [1]
which doesn't necessarily correspond to the height of the glyphs.  Thus
we get a situation where glyphs appear slightly above and below than the
line on which their drawn.  You might want that for typesetting
proportional fonts, but it doesn't make much sense in a monospace,
terminal-like application.

For calculating the global glyph height, the FreeType documentation says
to use the difference between the ascender and descender metrics.
Indeed, that is what urxvt [2] and xterm [3] do, and that is what is
implemented in this patch.

This change may cause alacritty to render a pixel or two more space
between lines depending on the font.  The spacing can be tweaked with
the `font.offset` setting at the expense of powerline separator
seamlessness.

[1] https://www.freetype.org/freetype2/docs/reference/ft2-base_interface.html#ft_facerec
[2] https://github.com/exg/rxvt-unicode/blob/rxvt-unicode-rel-9.22/src/rxvtfont.C#L1244
[3] https://github.com/Maximus5/xterm/blob/73ac00b39ec6f4af3b43908996a4bc73410a47df/src/fontutils.c#L833